Fixed ConcurrentModification that happens in async
User strings updated in an asynchronous block now get marked for
changing. the marked strings get updated on the next available game tick
so that it is in time with the main thread.

User strings are now stored in a hashmap to improve performance and
usablility.
Removed the ability to store strings in lists.
817a5e31cd9a2cdbcf1a4336ac8bbab35917f5b1

Added trig string functions thanks to @FalseHonesty
97e6c3b7f4b474dd69e0c3a0dfc7146d11769764

Download:
http://ct.kerbybit.com/download/versions/5.0/ChatTriggers_1.8.9_5.0b91.jar